backend/requirements.txt
Django==1.11.4
django-webpack-loader==0.4.1
channels==1.1.2
django-model-utils==2.6.1
libthumbor==1.3.2
djangorestframework==3.6.2
django-treebeard==4.1.0
celery[redis]==4.1.0
django-celery-monitor==1.1.2
psycopg2==2.7.3.1
